Added put, delete, get all players.
